fix: fix duplicate value mishandling
The previous search strategy of using JQ to access a path, copying the
value, and grepping for it didn't work because different keys might have
the same value, leading to multiple match results.

This new strategy uses `vim-jsonpath` to extract the actual position
within the JSON file. It's a lot slower, but it's also much more
consistent.